+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed reconnections not triggering `audioVideoWasDemotedFromPrimaryMeeting`.
+
 ## [3.21.1] - 2024-03-28
 
 ### Added